requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
alexwlchan-chives==23
    # via -r requirements.in
anyio==4.12.1
    # via httpx
brotli==1.2.0
    # via yt-dlp
certifi==2026.2.25
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
    #   yt-dlp
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
curl-cffi==0.14.0
    # via yt-dlp
gallery-dl==1.31.7
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
hyperlink==21.0.0
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
mutagen==1.47.0
    # via yt-dlp
pillow==12.1.1
    # via alexwlchan-chives
pycparser==3.0
    # via cffi
pycryptodomex==3.23.0
    # via yt-dlp
requests==2.32.5
    # via
    #   gallery-dl
    #   yt-dlp
urllib3==2.6.3
    # via
    #   requests
    #   yt-dlp
websockets==16.0
    # via yt-dlp
yt-dlp==2026.2.21
    # via -r requirements.in
yt-dlp-ejs==0.5.0
    # via yt-dlp
